# ConsoleAppClientGen/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My School API",
    "description": "A simple school ASP.NET Core Web API",
    "termsOfService": "http://contoso.com",
    "contact": {
      "name": "Shayne Boyer",
      "url": "https://twitter.com/spboyer"
    },
    "license": {
      "name": "Use under LICX",
      "url": "https://example.com/license"
    },
    "version": "v1"
  },
  "paths": {
    "/Classroom": {
      "get": {
        "tags": [
          "Classroom"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Classroom"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Classroom/{id}": {
      "get": {
        "tags": [
          "Classroom"
        ],
        "summary": "Get a classroom",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the classroom",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Classroom"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Classroom"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ClassroomSync": {
      "get": {
        "tags": [
          "ClassroomSync"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ClassroomSync"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ClassroomSync/{id}": {
      "get": {
        "tags": [
          "ClassroomSync"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "ClassroomSync"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ClassroomSync"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ClassroomTest": {
      "get": {
        "tags": [
          "ClassroomTest"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ClassroomTest"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ClassroomTest/{id}": {
      "get": {
        "tags": [
          "ClassroomTest"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "ClassroomTest"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Classroom"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ClassroomTest"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Student/class/{classroomid}": {
      "get": {
        "tags": [
          "Student"
        ],
        "parameters": [
          {
            "name": "classroomid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Teacher/class/{classroomid}": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "parameters": [
          {
            "name": "classroomid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Classroom": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "classroomID": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string"
          },
          "floor": {
            "type": "integer",
            "format": "int32"
          },
          "corridor": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}